uvision_project_HDMHD/PythonQTinterface/requirements.txt
numpy==1.26.0
pandas==2.1.1
pyserial==3.5
PySide6==6.5.3
PySide6-Addons==6.5.3
PySide6-Essentials==6.5.3
python-dateutil==2.8.2
pytz==2023.3.post1
shiboken6==6.5.3
six==1.16.0
tzdata==2023.3
